# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalDecision": {
        "enum": [
          "approved",
          "denied"
        ],
        "title": "ApprovalDecision",
        "type": "string"
      },
      "ApprovalDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "approver": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Compatibility assertion only; the server derives the effective approver from the authenticated operator identity",
            "title": "Approver"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "expected_digest": {
            "title": "Expected Digest",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "expected_digest"
        ],
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RXPVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "ledger": {
            "additionalProperties": true,
            "title": "Ledger",
            "type": "object"
          }
        },
        "required": [
          "ledger"
        ],
        "title": "RXPVerifyRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Evidence-gated, deterministic control plane for multi-agent embodied-AI research. The bundled EgoLite run is explicitly synthetic.",
    "title": "EgoAgentOS ResearchOps API",
    "version": "0.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals/{approval_id}/decision": {
      "post": {
        "operationId": "approval_decision_api_v1_approvals__approval_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approval Decision Api V1 Approvals  Approval Id  Decision Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approval Decision",
        "tags": [
          "approval"
        ]
      }
    },
...
    "/api/v1/demo/reset": {
      "post": {
        "operationId": "reset_demo_api_v1_demo_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DemoResetRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reset Demo Api V1 Demo Reset Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Demo",
        "tags": [
          "demo"
        ]
      }
    },
    "/api/v1/expert-runs": {
      "post": {
        "operationId": "create_expert_run_api_v1_expert_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpertRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Expert Run Api V1 Expert Runs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Expert Run",
        "tags": [
          "research-os"
        ]
      }
    },
...
    "/api/v1/expert-runs/{run_id}": {
      "get": {
        "operationId": "expert_run_api_v1_expert_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Expert Run Api V1 Expert Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Expert Run",
        "tags": [
          "research-os"
        ]
      }
    },
...
    "/api/v1/rxp/verify": {
      "post": {
        "operationId": "rxp_verify_api_v1_rxp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RXPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rxp Verify Api V1 Rxp Verify Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rxp Verify",
        "tags": [
          "rxp"
        ]
      }
    },
    "/api/v1/skill-invocations/{invocation_id}": {
      "get": {
        "operationId": "skill_trace_api_v1_skill_invocations__invocation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "invocation_id",
            "required": true,
            "schema": {
              "title": "Invocation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Skill Trace Api V1 Skill Invocations  Invocation Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skill Trace",
        "tags": [
          "skills"
        ]
      }
    },
...
  }
}